EmulatorView has to hold the disc state, since emulator gets recreated when changing models
